# contracts/housing-service.openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HousingService",
    "version": "1.0"
  },
  "paths": {
...
    "/api/allocations/candidate-rooms": {
      "get": {
        "tags": [
          "Allocations"
        ],
        "parameters": [
          {
            "name": "housingRequestId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "housingGroupId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateRoomDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateRoomDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateRoomDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/allocations": {
      "post": {
        "tags": [
          "Allocations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAllocationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAllocationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAllocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Allocations"
        ],
        "parameters": [
          {
            "name": "buildingId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "roomId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/allocations/{id}": {
      "get": {
        "tags": [
          "Allocations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/housing-groups": {
      "post": {
        "tags": [
          "HousingGroups"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHousingGroupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHousingGroupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHousingGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "HousingGroups"
        ],
        "parameters": [
          {
            "name": "housingCycleId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/housing-groups/join": {
      "post": {
        "tags": [
          "HousingGroups"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinHousingGroupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinHousingGroupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/JoinHousingGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInvitationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInvitationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInvitationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/housing-groups/invitations/{invitationId}/respond": {
      "post": {
        "tags": [
          "HousingGroups"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInvitationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInvitationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/housing-groups/{id}": {
      "get": {
        "tags": [
          "HousingGroups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingGroupDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/housing-groups/{id}/members/{studentId}/remove": {
      "post": {
        "tags": [
          "HousingGroups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AllocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "housingRequestId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "housingGroupId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "roomId": {
            "type": "integer",
            "format": "int32"
          },
          "roomNumber": {
            "type": "string",
            "nullable": true
          },
          "buildingId": {
            "type": "integer",
            "format": "int32"
          },
          "buildingName": {
            "type": "string",
            "nullable": true
          },
          "occupantStudentIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "allocatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "vacatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AllocationDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllocationDto"
            },
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "CandidateRoomDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "integer",
            "format": "int32"
          },
          "roomNumber": {
            "type": "string",
            "nullable": true
          },
          "floor": {
            "type": "integer",
            "format": "int32"
          },
          "buildingId": {
            "type": "integer",
            "format": "int32"
          },
          "buildingName": {
            "type": "string",
            "nullable": true
          },
          "remainingCapacity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateAllocationDto": {
        "type": "object",
        "properties": {
          "housingRequestId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "housingGroupId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "roomId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateHousingGroupDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GroupInvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "invitedStudentId": {
            "type": "string",
            "nullable": true
          },
          "invitedStudentName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/InvitationStatus"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "HousingGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "leaderId": {
            "type": "string",
            "nullable": true
          },
          "housingCycleId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/HousingGroupStatus"
          },
          "maxMembers": {
            "type": "integer",
            "format": "int32"
          },
          "memberStudentIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "pendingInvitations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupInvitationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HousingGroupDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HousingGroupDto"
            },
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "HousingGroupStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "InvitationStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "JoinHousingGroupDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RespondToInvitationDto": {
        "type": "object",
        "properties": {
          "approve": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}